pyproject.toml
[project]
name = "browser-use-benchmark"
version = "0.1.0"
description = "Open source benchmark for browser-use"
requires-python = ">=3.11"
dependencies = [
    "browser-use==0.11.5",
    "certifi",
    "cryptography",
    "httpx",
    "python-dotenv",
    "spider-browser",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
plots = ["matplotlib", "numpy"]

[tool.hatch.build.targets.wheel]
packages = ["."]
